Handle ScenarioGUI Activity config and starting in separate class ScenarioActivityConfigGUI - non function, reimplement later
